# openapi/aca-openapi-v2.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "ACA-Upload-API-v2",
    "description" : "Microservice to manage PagoPA GPD Upload",
    "termsOfService" : "https://www.pagopa.gov.it/",
    "version" : "0.5.2"
  },
  "servers" : [ {
    "url" : "https://{host}{basePath}",
    "variables" : {
      "basePath" : {
        "default" : "/upload/aca/debt-positions-service/v2",
        "enum" : [ "/upload/aca/debt-positions-service/v2" ]
      },
      "host" : {
        "default" : "api.platform.pagopa.it",
        "enum" : [ "api.uat.platform.pagopa.it", "api.platform.pagopa.it" ]
      }
    }
  } ],
  "paths" : {
...
    "/brokers/{broker-code}/organizations/{organization-fiscal-code}/debtpositions/file/{file-id}/status" : {
      "get" : {
        "tags" : [ "Massive operation observability APIs - v2" ],
        "summary" : "Returns the debt positions upload status.",
        "operationId" : "get-debt-positions-upload-status",
        "parameters" : [ {
          "name" : "broker-code",
          "in" : "path",
          "description" : "The broker code",
          "required" : true,
          "schema" : {
            "minLength" : 1,
            "type" : "string"
          }
        }, {
          "name" : "organization-fiscal-code",
          "in" : "path",
          "description" : "The organization fiscal code",
          "required" : true,
          "schema" : {
            "minLength" : 1,
            "type" : "string"
          }
        }, {
          "name" : "file-id",
          "in" : "path",
          "description" : "The unique identifier for file upload",
          "required" : true,
          "schema" : {
            "minLength" : 1,
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Upload found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UploadStatusDTO"
                }
              }
            }
          },
          "400" : {
            "description" : "Malformed request.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "401" : {
            "description" : "Wrong or missing function key.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          },
          "404" : {
            "description" : "Upload not found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "429" : {
            "description" : "Too many requests.",
            "content" : {
              "text/json" : { }
            }
          },
          "500" : {
            "description" : "Service unavailable.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            }
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "OperationStatus" : {
        "type" : "string",
        "enum" : [ "IN_PROGRESS", "COMPLETED", "COMPLETED_WITH_WARNINGS", "COMPLETED_UNSUCCESSFULLY" ]
      },
      "ProblemJson" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string",
            "description" : "A short, summary of the problem type. Written in english and readable for engineers (usually not suited for non technical stakeholders and not localized); example: Service Unavailable"
          },
          "status" : {
            "type" : "integer",
            "description" : "The HTTP status code generated by the origin server for this occurrence of the problem.",
            "format" : "int32",
            "example" : 400
          },
          "detail" : {
            "type" : "string",
            "description" : "A human readable explanation specific to this occurrence of the problem.",
            "example" : "There was an error processing the request"
          }
        },
        "description" : "Object returned as response in case of an error."
      },
...
      "UploadStatusDTO" : {
        "type" : "object",
        "properties" : {
          "fileId" : {
            "type" : "string"
          },
          "processedItem" : {
            "type" : "integer",
            "format" : "int32"
          },
          "submittedItem" : {
            "type" : "integer",
            "format" : "int32"
          },
          "startTime" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2024-10-08T14:55:16.302Z"
          },
          "operationStatus" : {
            "$ref" : "#/components/schemas/OperationStatus"
          }
        }
      }
    },
    "securitySchemes" : {
      "ApiKey" : {
        "type" : "apiKey",
        "name" : "ApiKey",
        "in" : "header"
      }
    }
  }
}